requirements.txt
streamlit
pandas
io
matplotlib.pyplot
seaborn
numpy
sklearn